+++ network/messagequeue.cpp
@@ -184,7 +184,7 @@ void	Messagequeue::push_out_textmsg(std::string const &str)
 		_msgout[_frontout].cltid = 0;
 		_msgout[_frontout].actid = 0;
 		_msgout[_frontout].ping = 0.0f;
-		_msgout[_frontout].pckt(str.length() + 1, str.c_str());
+		_msgout[_frontout].pckt((unsigned int)str.length() + 1, str.c_str());
 		_frontout = (_frontout + 1) % _size;
 	}
 }
@@ -197,7 +197,7 @@ void	Messagequeue::push_out_textmsg(int cltid, std::string const &str)
 		_msgout[_frontout].cltid = cltid;
 		_msgout[_frontout].actid = 0;
 		_msgout[_frontout].ping = 0.0f;
-		_msgout[_frontout].pckt(str.length() + 1, str.c_str());
+		_msgout[_frontout].pckt((unsigned int)str.length() + 1, str.c_str());
 		_frontout = (_frontout + 1) % _size;
 	}
 }
